+++ api/src/main/java/com/gpudb/protocol/AdminRebalanceRequest.java
@@ -103,8 +103,7 @@ public class AdminRebalanceRequest implements IndexedRecord {
      * com.gpudb.protocol.AdminRebalanceRequest.Options#COMPACT_AFTER_REBALANCE
      * COMPACT_AFTER_REBALANCE}: Perform compaction of deleted records once the
      * rebalance completes, to reclaim memory and disk space. Default is true,
-     * unless {add_labels}@{key of options repair_incorrectly_sharded_data} is
-     * set to {@code true}.
+     * unless {@code repair_incorrectly_sharded_data} is set to {@code true}.
      * Supported values:
      * <ul>
      *         <li> {@link
@@ -214,9 +213,8 @@ public class AdminRebalanceRequest implements IndexedRecord {
 
         /**
          * Perform compaction of deleted records once the rebalance completes,
-         * to reclaim memory and disk space. Default is true, unless
-         * {add_labels}@{key of options repair_incorrectly_sharded_data} is set
-         * to {@code true}.
+         * to reclaim memory and disk space. Default is true, unless {@code
+         * repair_incorrectly_sharded_data} is set to {@code true}.
          * Supported values:
          * <ul>
          *         <li> {@link
@@ -348,8 +346,7 @@ public class AdminRebalanceRequest implements IndexedRecord {
      *                 com.gpudb.protocol.AdminRebalanceRequest.Options#COMPACT_AFTER_REBALANCE
      *                 COMPACT_AFTER_REBALANCE}: Perform compaction of deleted
      *                 records once the rebalance completes, to reclaim memory
-     *                 and disk space. Default is true, unless
-     *                 {add_labels}@{key of options
+     *                 and disk space. Default is true, unless {@code
      *                 repair_incorrectly_sharded_data} is set to {@code true}.
      *                 Supported values:
      *                 <ul>
@@ -470,8 +467,8 @@ public class AdminRebalanceRequest implements IndexedRecord {
      *         com.gpudb.protocol.AdminRebalanceRequest.Options#COMPACT_AFTER_REBALANCE
      *         COMPACT_AFTER_REBALANCE}: Perform compaction of deleted records
      *         once the rebalance completes, to reclaim memory and disk space.
-     *         Default is true, unless {add_labels}@{key of options
-     *         repair_incorrectly_sharded_data} is set to {@code true}.
+     *         Default is true, unless {@code repair_incorrectly_sharded_data}
+     *         is set to {@code true}.
      *         Supported values:
      *         <ul>
      *                 <li> {@link
@@ -591,8 +588,7 @@ public class AdminRebalanceRequest implements IndexedRecord {
      *                 com.gpudb.protocol.AdminRebalanceRequest.Options#COMPACT_AFTER_REBALANCE
      *                 COMPACT_AFTER_REBALANCE}: Perform compaction of deleted
      *                 records once the rebalance completes, to reclaim memory
-     *                 and disk space. Default is true, unless
-     *                 {add_labels}@{key of options
+     *                 and disk space. Default is true, unless {@code
      *                 repair_incorrectly_sharded_data} is set to {@code true}.
      *                 Supported values:
      *                 <ul>